Refactor Header component to use new file structure
